pyproject.toml
[project]
name = "src-bnb-agentic-social-platform-promo-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # Core AI & OpenAI
    "openai>=1.52.0",
    "tiktoken>=0.7.0",
    
    # Web & API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.12.0",
    
    # Data & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Environment & Config
    "python-dotenv>=1.0.0",
    
    # Utilities
    "tenacity>=8.2.0",
    "pip",
    "pyyaml",]
